# fix(perf): simplify eventstore queries by removing or in projection handlers