--- a/internal/api/oidc/key.go
+++ b/internal/api/oidc/key.go
@@ -419,13 +419,14 @@ func (o *OPStorage) getMaxKeySequence(ctx context.Context) (float64, error) {
 			AwaitOpenTransactions().
 			AllowTimeTravel().
 			AddQuery().
-			AggregateTypes(keypair.AggregateType).
+			AggregateTypes(
+				keypair.AggregateType,
+				instance.AggregateType,
+			).
 			EventTypes(
 				keypair.AddedEventType,
+				instance.InstanceRemovedEventType,
 			).
-			Or().
-			AggregateTypes(instance.AggregateType).
-			EventTypes(instance.InstanceRemovedEventType).
 			Builder(),
 	)
 }